+++ src/components/th/th-styles.jsx
@@ -2,10 +2,11 @@ import { createStyleSheet } from 'jss-theme-reactor';
 import styleUtils from '../table/style-utilities';
 
 export default createStyleSheet('Th', (theme) => {
-  const { palette, typography } = theme;
+  const { palette, typography, mixins } = theme;
 
   return {
     th: {
+      ...mixins.basicBoxSizing,
       ...styleUtils.flexItem(),
       ...styleUtils.sizes(),
       ...styleUtils.modifiers(palette),