+++ src/components/tr/tr-styles.jsx
@@ -3,12 +3,14 @@ import styleUtils from '../table/style-utilities';
 import color from '../../styles/color';
 
 export default createStyleSheet('Tr', (theme) => {
-  const { palette } = theme;
+  const { palette, mixins } = theme;
 
   return {
     tr: {
+      ...mixins.basicBoxSizing,
       ...styleUtils.flexRow(),
       ...styleUtils.sizes(),
+      alignItems: 'center',
       borderLeft: '1px solid transparent',
       borderRight: '1px solid transparent',