+++ src/styles/theme.js
@@ -1,17 +1,17 @@
 // import shadows from './shadows';
 import transitions from './transitions';
 import createTypography from './typography';
-// import createBreakpoints from './breakpoints';
+import createBreakpoints from './breakpoints';
 import { createPalette } from './palette';
 // import zIndex from './zIndex';
-// import createMixins from './mixins';
+import createMixins from './mixins';
 // import spacing from './spacing';
 
 export function createTheme(config = {}) {
   const {
     palette = createPalette(),
-    breakpoints = {},
-    mixins = {},
+    breakpoints = createBreakpoints(),
+    mixins = createMixins(breakpoints),
     typography = createTypography(),
     ...more
   } = config;
@@ -20,7 +20,7 @@ export function createTheme(config = {}) {
     palette,
     typography,
     mixins,
-    breakpoints, /* shadows, spacing, zIndex */
+    breakpoints /* shadows, spacing, zIndex */,
     transitions,
     ...more,
   };